cleanup global function by either moving them to class or merge the .h in to its .cc
