--- RecoTracker/DebugTools/plugins/GluedDetFromDetUnit.h
@@ -1,6 +1,0 @@
-#include "DataFormats/DetId/interface/DetId.h"
-
-inline DetId gluedId(const DetId& du) {
-  unsigned int mask = ~3;  // mask the last two bits
-  return DetId(du.rawId() & mask);
-}
